requirements_export2.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
aiohttp=3.9.1=pypi_0
aioice=0.9.0=pypi_0
aiortc=1.6.0=pypi_0
aiosignal=1.3.1=pypi_0
altair=4.2.2=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
av=11.0.0=pypi_0
blas=1.0=mkl
blinker=1.7.0=pypi_0
brotli-python=1.0.9=py39h5a03fae_7
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.11.17=hbcca054_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pyhd8ed1ab_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
chumpy=0.70=pypi_0
click=8.1.7=pypi_0
contourpy=1.2.0=pypi_0
cryptography=41.0.7=pypi_0
cudatoolkit=11.3.1=h9edb442_10
cycler=0.12.1=pypi_0
dnspython=2.4.2=pypi_0
easydict=1.13=pypi_0
einops=0.7.0=pypi_0
entrypoints=0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
face-alignment=1.4.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.19.1=pypi_0
filterpy=1.4.5=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.46.0=pypi_0
freetype=2.10.4=h0708190_1
frozenlist=1.4.0=pypi_0
fsspec=2023.12.1=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gast=0.4.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.25.1=pypi_0
google-auth-oauthlib=1.1.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.60.0=pypi_0
h5py=3.1.0=pypi_0
hf-xet=1.1.9=pypi_0
huggingface-hub=0.34.4=pypi_0
idna=3.6=pyhd8ed1ab_0
ifaddr=0.2.0=pypi_0
imageio=2.33.0=pypi_0
importlib-metadata=7.0.0=pypi_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.10=pypi_0
jinja2=3.1.2=pypi_0
jpeg=9e=h166bdaf_1
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.11.2=pypi_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.7.0=pypi_0
lame=3.100=h7f98852_1001
lazy-loader=0.3=pypi_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.6=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=hf544144_3
libwebp-base=1.2.2=h7f98852_1
lightning-utilities=0.10.0=pypi_0
llvmlite=0.41.1=pypi_0
loguru=0.7.2=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.5.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
mdurl=0.1.2=pypi_0
mediapipe=0.10.8=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7e14d7c_0
mkl_fft=1.3.1=py39h0c7bc48_1
mkl_random=1.2.2=py39hde0f152_0
ml-dtypes=0.2.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.6=he412f7d_0
networkx=3.2.1=pypi_0
ninja=1.11.1.1=pypi_0
numba=0.58.1=pypi_0
numpy=1.23.0=pypi_0
oauthlib=3.2.2=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-contrib-python=4.8.1.78=pypi_0
opencv-python=4.8.1.78=pypi_0
opencv-python-headless=4.6.0.66=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openpyxl=3.1.2=pypi_0
openssl=3.0.12=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pypi_0
pandas=1.4.4=pypi_0
pillow=9.5.0=pypi_0
pip=23.3.1=py39h06a4308_0
portalocker=2.8.2=pypi_0
protobuf=3.20.3=pypi_0
pyarrow=14.0.1=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydeck=0.8.1b0=pypi_0
pyee=11.1.0=pypi_0
pygments=2.17.2=pypi_0
pylibsrtp=0.9.1=pypi_0
pympler=1.0.1=pypi_0
pyopenssl=23.3.0=pypi_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python_abi=3.9=2_cp39
pytorch=1.12.1=py3.9_cuda11.3_cudnn8.3.2_0
pytorch-lightning=2.1.2=pypi_0
pytorch-mutex=1.0=cuda
pytorch3d=0.7.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.32.0=pypi_0
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
retinaface=1.1.1=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.13.2=pypi_0
rsa=4.9=pypi_0
safetensors=0.6.2=pypi_0
scikit-image=0.22.0=pypi_0
scipy=1.11.4=pypi_0
semver=3.0.2=pypi_0
setuptools=68.0.0=py39h06a4308_0
six=1.15.0=pypi_0
smmap=5.0.1=pypi_0
sounddevice=0.4.6=pypi_0
sqlite=3.41.2=h5eee18b_0
streamlit=1.10.0=pypi_0
streamlit-webrtc=0.42.0=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.20.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=1.1.0=pypi_0
tifffile=2023.9.26=pypi_0
timm=1.0.19=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torchaudio=0.12.1=py39_cu113
torchfile=0.1.0=pypi_0
torchmetrics=1.2.1=pypi_0
torchvision=0.13.1=py39_cu113
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=5.2=pypi_0
urllib3=2.1.0=pyhd8ed1ab_0
validators=0.22.0=pypi_0
vit-pytorch=1.6.4=pypi_0
watchdog=3.0.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py39h06a4308_0
wrapt=1.12.1=pypi_0
xz=5.4.2=h5eee18b_0
yacs=0.1.8=pypi_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.0=ha95c52a_0
